+++ .gitignore
@@ -1,0 +1,2 @@
+# INDICA AO GIT PARA NAO ENVIAR AS PASTAS PRO REPOSITÓRIO
+node_modules
